chore: #286 userService util seperating
